+++ main.rb
@@ -3,9 +3,9 @@ require './app'
 def main
   app = App.new
   app.run
-
+  
   at_exit do
-    app.preserve_all
+  app.preserve_all
   end
 end
