+++ rentals.json
@@ -1,1 +1,1 @@
-[{"date":"23","book_index":1,"person_index":0},{"date":"45","book_index":0,"person_index":0}]
+[{"date":"2022-20-20","book_index":1,"person_index":0}]
